Avoid assertion failurein pq_getbyte and pq_getbytes.
After the commit 'Be more careful to not lose sync in the FE/BE protocol',
PostgreSQL requires an assertion flags in pq_getbyte and pq_getbytes,
but pg_bulkload call these without setting a flag. The assertion
failure happens on PostgreSQL 9.4.1, 9.3.6, 9.2.10, 9.1.15, 9.0.19.
With this patch, the wrapper function for these API are introduced
in pg_bulkload side and set a flag if needed.