Fix writing of checksum for blocks around the segment boundary.
In direct write mode, pg_bulkload loads data into a fixed number of
private buffers and once they are full, writes those buffers to the
actual relation file.  Each page's checksum is set before writing to
the file, during which it must pass the block number for checksum
calculation.  Due to a bug in how block number is calculated, it
calculated the wrong checksum in certain cases, especially, when a
segment of the relation is about to get full.