Fix crash due to improper memory context handling in some cases
AsyncSource mechanism which uses a separate thread to read data can
sometimes try to manipulate an address that may be invalid because of
memory context reset or delete by the main thread.

Fix that by implementing a dedicated memory context for AsyncSource,
which is only manipulated in the reading thread.

Fixes #27.