Prevent CSVParserRead() from requesting too much memory
Currently, its record buffer and field buffer each start at 1MB
and later double it multiple times.  If the doubling occurs too many
times, at some point, we reach the maximum limit on request size to
palloc, which causes the error: "invalid memory alloc request size
1074789376"

Fix that by putting a maximum limit on the request size = 16MB.  We
can increase that later if it is found that it's too low.

Fixes #29.