Error out sooner upon finding an unterminated quoted field
Currently, if there is an unquoted field in a CSV file, we don't
error out soon enough.  For example, if we find a newline before we
find the terminating quote, it's safe to assume that we've encountered
an unquoted field and error out right away.  Currently, we don't quit
until the SourceRead finishes reading the whole file (which could be
quite long), which is entirely wasteful.

Minor comment cleanup as well.