Fix build against PG12
Non-trivial changes are needed for "parallel"/"multi-process" mode
to work, because PG, as of v12, doesn't allow opening a relation
without holding a lock.  The fix involves the frontend-facing
process to open the target relation with an AccessShareLock and
then relinquish it for the writer process to be able to acquire
an AccessExclusiveLock.  The fronend-facing process holds the
relation open without holding any lock which was the case even
before this patch.

Few other fixes due to PG-side API changes, along with the usual
copying of nbtsort.c.

There is bunch of churn in regression test outputs due to the
default value of extra_float_digits GUC parameter having changed
in core PG and some error messages.  Adjust alternate expected
output files accordingly.