PostgreSQL 13 support - phase 3
Corrected related to regression test.

In the case of PG13, there was a change to the binary cast in the process of getting the text data of the user-function.
But, pg_bulkload does not respond to this point, so there is a problem that text columns are not converted properly.
For related details, please refer to write_bin.sql.

We need to fix this, but it's not a frequently used feature, so we'll support it later.