add document to specify returning data type.
This patch handles #93 issue. From PG13, INPUT=FUNCTION may
fail if using SETOF RECORD because the assingment cast from
char/varchar to cstring type which pg_bulkload use internally.

So, this patch add the workaround way in the document. This
describes it's necessary to specify returning data type of functions,
which is also described in the PG13 document.