Fix duplication handling for large data set(#148)
BTReaderInit() assumed that the case (itup->t_tid).ip_posid = 0 was anomalous when searching for the leftmost leaf page of a b-tree, but this is an old condition and has been removed.

With the commit (postgres/postgres@8224de4) to support the INCLUDE clause in indexes from PostgreSQL v11, new pivot tuples (e.g. b-tree root and intermediate nodes) were added, which could cause ip_posid to be 0.

Because they had forgotten to address this, errors sometimes occurred when running the process on B-trees that held some large data with pivot tuples.